Resolve SessionDeleter Compiler Warning
It looks like the lack of a constructor is at the root of the compiler
warnings when you try to alias a class.

To stop the error, I have made a small modification to the session
deleter class so that it could have a constructor.

Change-Id: Iad4614d80dc317b20f01930019096fb388b74ed9
(cherry picked from commit 7426cfa1f0b3fe54803f9f489096e7b3777d99da)